Fix Update playlist functionality
Before, a User was able to delete the
remaining song from a playlist. Fixed
functionality to disable deleting single
song from playlist. User will have to
delete entire playlist if they want to
remove the only song. Currently, user
can't have an empty Playlist and add
songs to that empty list. That is a
plus feature in a later version.