Fix jwt bug and upgrade GraphQL server
Bug with jwt auth flow had to do with how we
were accessing a user found in the database.
User object was inside an array, was trying
to access user ID property when actual
returned value was an array with a user
inside. GraphQL server has been upgraded to
ApolloServer v2. Must now clean up and
refactor server files.